give kafkaviewer it's own little docker-compose again
Idea now is to run this docker-compose file with the main game-on one by
specifying multiple -f arguments to docker compose.

This does allow one docker-compose to make links to services defined in
the other one.

I've tested this with success, although I'm not entirely able to explain
how docker-compose is able to find the env_file configured, but testing
suggests plainly that it able to do so.

Signed-off-by: Ozzy Osborne <ozzy@ca.ibm.com>